messenger-service/requirements.txt
django==3.2.16
django-environ
mysqlclient
channels==3.0.1
pyopenssl==21.0.0
cryptography==36.0.1
django-db-views
djangorestframework == 3.11.0
coreapi
requests_toolbelt
py_eureka_client
gunicorn
uvicorn
channels-redis==2.4.2
daphne==2.5.0
djangorestframework-jwt
websockets
pyfcm==1.5.4
opengraph_parse
pyforkurento
django-storages[google]
flake8